+++ nengo/__init__.py
@@ -29,6 +29,7 @@ from .probe import Probe
 from .rc import rc, RC_DEFAULTS
 from .simulator import Simulator
 from .synapses import Alpha, LinearFilter, Lowpass, Triangle
+from .transforms import Convolution
 from .utils.logging import log
 from . import dists, exceptions, networks, presets, processes, spa, utils